Adds basic UI and message injection on Twitch.
This is the first real commit to the project. I add prettier for a formatter, Some React components for the UI, and most notably Twitch chat injection. This allows for proof of concept, where arbitrary messages are passed to the same function as the web socket, allowing for a native experience.