Create LICENCE
Adds the Business Source License. This license was chosen because it was effectively a source available license that would become open source after a set period of time. Source available is a significant factor for Chrome extensions, as they can have a lot of control over the browser, and verifying that the code is not malicious is critical. Additionally, the paid service is intended to fund the extension's development, so keeping it source available while I am working on the project feels best.